ease-of-use for k6
