--- load/participants
@@ -1,1 +1,0 @@
-GET https://hcapparticipants.dev.freshworks.club/api/v1/version
